Clean code fixes CRUD created

+++ wizyter/src/main/java/com/szymon/wizyter/Services/PatientService.java
@@ -39,7 +39,12 @@ public class PatientService {
         });
     }
 
-    public void deletePatient(Long id) {
-        patientRepository.deleteById(id);
+    public boolean deletePatientIfExists(Long id) {
+        if (patientRepository.existsById(id)) {
+            patientRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
 }
